+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # [NEXT]
+## Changed
+- Only show recent characters from the same fileset. (#173)
+- Updates to Gitmoji and FontAwesome.
+
 ## Added
 - Support for [tofi](https://github.com/philj56/tofi/), [bemenu](https://github.com/Cloudef/bemenu) and [wmenu](https://git.sr.ht/~adnano/wmenu) as selectors
-
-## Changed
-- Updates to Gitmoji and FontAwesome.
 
 # [6.2.0]
 ## Added